+++ examples/postgres/src/main/java/com/github/braisdom/objsql/example/TransactionalExample.java
@@ -13,9 +13,9 @@ public class TransactionalExample {
     public static void createNormally() {
         Order order = new Order()
                 .setNo("202000001")
-                .setMemberId(3)
-                .setAmount(3.5d)
-                .setQuantity(100.3d)
+                .setMemberId((long)3)
+                .setAmount(3.5f)
+                .setQuantity(100.3f)
                 .setSalesAt(Timestamp.valueOf("2020-05-01 09:30:00"));
 
         ExecutorService executorService = Executors.newFixedThreadPool(4);
