+++ examples/postgres/src/main/java/com/github/braisdom/objsql/example/domains/Order.java
@@ -14,9 +14,9 @@ import java.util.List;
 @DomainModel
 public class Order {
     private String no;
-    private Integer memberId;
-    private Double amount;
-    private Double quantity;
+    private Long memberId;
+    private Float amount;
+    private Float quantity;
 
     @Column(transition = SqlDateTimeTransition.class)
     private Timestamp salesAt;
